requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.109
botocore==1.20.109
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.4.5
django-timezone-field==4.2.1
idna==2.10
jmespath==0.10.0
kombu==5.1.0
mysql-connector-python==8.0.21
mysqlclient==2.0.3
pip-autoremove==0.9.1
prompt-toolkit==3.0.19
protobuf==3.17.3
pycodestyle==2.7.0
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
